feat: migrate the `/oauth` & `/oauth-result` pages to app router
