+++ src/app/oauth-result/types.ts
@@ -1,0 +1,14 @@
+import { PlatformName } from "@guildxyz/types"
+
+export type OAuthResultParams =
+  | {
+      status: "success"
+      platform: PlatformName
+      path: string
+    }
+  | {
+      status: "error"
+      message: string
+      platform?: PlatformName
+      path?: string
+    }
