+++ src/components/[guild]/JoinModal/hooks/useConnectPlatform.ts
@@ -2,6 +2,7 @@ import { usePostHogContext } from "@/components/Providers/PostHogProvider"
 import { platformMergeAlertAtom } from "@/components/Providers/atoms"
 import { useErrorToast } from "@/components/ui/hooks/useErrorToast"
 import { useToast } from "@/components/ui/hooks/useToast"
+import { OAuthResultParams } from "app/oauth-result/types"
 import useUser from "components/[guild]/hooks/useUser"
 import { env } from "env"
 import { useFetcherWithSign } from "hooks/useFetcherWithSign"
@@ -10,7 +11,6 @@ import usePopupWindow from "hooks/usePopupWindow"
 import useSubmit, { SignedValidation, useSubmitWithSign } from "hooks/useSubmit"
 import { UseSubmitOptions } from "hooks/useSubmit/types"
 import { useSetAtom } from "jotai"
-import { OAuthResultParams } from "pages/oauth-result"
 import { useCallback, useMemo } from "react"
 import rewards from "rewards"
 import useSWR from "swr"
